// packages/shared/src/enums/StoryCategory.ts
export enum StoryCategory {
  BUSINESS_AND_FINANCE = "business-and-finance",
  DIGITAL_GRAPHICS = "digital-graphics",
  DIY = "diy",
  ENTERTAINMENT = "entertainment",
  GAMING = "gaming",
  HEALTH_AND_WELLNESS = "health-and-wellness",
  LEARNING = "learning",
  LIFESTYLE = "lifestyle",
  MUSIC = "music",
  NEWS = "news",
  PROGRAMMING = "programming",
  SCIENCE_AND_TECHNOLOGY = "science-and-technology",
  SPORTS = "sports",
  TRAVEL = "travel"
}